+++ cmd/main.go
@@ -7,7 +7,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/eminetto/clean-architecture-go-v2/infra/repository"
+	"github.com/eminetto/clean-architecture-go-v2/driver/repository"
 	"github.com/eminetto/clean-architecture-go-v2/usecase/book"
 
 	"github.com/eminetto/clean-architecture-go-v2/config"
