+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
     image: mysql:5.7
     command: --default-authentication-plugin=mysql_native_password --init-file /data/application/init.sql
     volumes:
-      - ./infra/init.sql:/data/application/init.sql
+      - ./ops/init.sql:/data/application/init.sql
     environment:
       MYSQL_ROOT_PASSWORD: clean_architecture_go_v2
       MYSQL_DATABASE: clean_architecture_go_v2
@@ -30,7 +30,7 @@ services:
     command:
       - --config.file=/etc/prometheus/prometheus.yml
     volumes:
-      - ./infra/prometheus/prometheus.yml:/etc/prometheus/prometheus.yml:ro
+      - ./ops/prometheus/prometheus.yml:/etc/prometheus/prometheus.yml:ro
     container_name: clean-architecture-go-v2-prometheus
     network_mode: "bridge"
   prometheus-pushgateway:
