+++ usecase/user/service_test.go
@@ -4,8 +4,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/eminetto/clean-architecture-go-v2/infra/repository"
-
 	"github.com/eminetto/clean-architecture-go-v2/entity"
 
 	"github.com/stretchr/testify/assert"
@@ -23,7 +21,7 @@ func newFixtureUser() *entity.User {
 }
 
 func Test_Create(t *testing.T) {
-	repo := repository.NewUserInmem()
+	repo := newInmem()
 	m := NewService(repo)
 	u := newFixtureUser()
 	_, err := m.CreateUser(u.Email, u.Password, u.FirstName, u.LastName)
@@ -33,7 +31,7 @@ func Test_Create(t *testing.T) {
 }
 
 func Test_SearchAndFind(t *testing.T) {
-	repo := repository.NewUserInmem()
+	repo := newInmem()
 	m := NewService(repo)
 	u1 := newFixtureUser()
 	u2 := newFixtureUser()
@@ -66,7 +64,7 @@ func Test_SearchAndFind(t *testing.T) {
 }
 
 func Test_Update(t *testing.T) {
-	repo := repository.NewUserInmem()
+	repo := newInmem()
 	m := NewService(repo)
 	u := newFixtureUser()
 	id, err := m.CreateUser(u.Email, u.Password, u.FirstName, u.LastName)
@@ -83,7 +81,7 @@ func Test_Update(t *testing.T) {
 }
 
 func TestDelete(t *testing.T) {
-	repo := repository.NewUserInmem()
+	repo := newInmem()
 	m := NewService(repo)
 	u1 := newFixtureUser()
 	u2 := newFixtureUser()
